requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
openai==1.3.5
python-dotenv>=0.5.1
pyyaml==5.4.1
